fix usage of wrong `abs()` function
- on device use `::fabs()` for double values
- on host use `std:abs()` for double values
- use `::fabsf()` for float values

+++ src/libPMacc/include/algorithms/math/floatMath/abs.tpp
@@ -41,7 +41,7 @@ struct Abs<float>
 
     HDINLINE float operator( )(float value)
     {
-        return ::fabs( value );
+        return ::fabsf( value );
     }
 };
